Breaking Change: Update to leverage build process using keys.json (#11)
* Adding keys.json to be ignored

* Adding key.js as part of the background script.

This file is generated during the multi-build process.  It is based on keys.json as opposed to common/keys.js

* Updating request to use api key constant

* Removing export of getKey from interface

* Updating Readme for more clarity

Adjusted the way the key.json works and added a bit more structure to the doc.

* Updating Readme to reflect `_dev` change

* Adding index file for localhost instructions

This is just boilerplate at this point.  This should be fleshed out with more useful information

* Adding conditional basic auth to request

# Conflicts:
#	src/common/api/_request/request.js

* Updating readme to indicate proper file location